Renaming a navigator unit test.
